+++ MatchdayPredictions.Api/DataAccess/Interfaces/ILeagueDataContext.cs
@@ -1,0 +1,11 @@
+using MatchdayPredictions.Api.Models.Api;
+
+namespace MatchdayPredictions.Api.DataAccess.Interfaces;
+
+public interface ILeagueDataContext
+{
+    Task<IEnumerable<League>> GetLeaguesAsync();
+
+    Task<League?> GetLeagueByIdAsync(int leagueId);
+}
+