+++ MatchdayPredictions.Api/DataAccess/Interfaces/IUserDataContext.cs
@@ -1,0 +1,13 @@
+using MatchdayPredictions.Api.Models.Api;
+
+namespace MatchdayPredictions.Api.DataAccess.Interfaces;
+
+public interface IUserDataContext
+{
+    Task CreateUserAsync(string username, string displayName, string email, string passwordHash);
+
+    Task<User?> GetUserByIdAsync(int userId);
+
+    Task<User?> GetUserByUsernameAsync(string username);
+}
+