+++ MatchdayPredictions.Api/Repositories/Interfaces/ILeagueRepository.cs
@@ -1,22 +1,14 @@
-﻿using MatchdayPredictions.Api.Models.Api;
+using MatchdayPredictions.Api.Models.Api;
 
-namespace MatchdayPredictions.Api.DataAccess.Repository;
+namespace MatchdayPredictions.Api.Repositories.Interfaces;
 
 /// <summary>
 /// Defines data operations for football leagues.
 /// </summary>
 public interface ILeagueRepository
 {
-    /// <summary>
-    /// Retrieves all leagues.
-    /// </summary>
-    /// <returns>A collection of all leagues.</returns>
     Task<IEnumerable<League>> GetAllAsync();
 
-    /// <summary>
-    /// Retrieves a specific league by its ID.
-    /// </summary>
-    /// <param name="leagueId">The league ID.</param>
-    /// <returns>The league if found, otherwise null.</returns>
     Task<League?> GetByIdAsync(int leagueId);
 }
+